Update README.md
In this project we tackle the classification problem using Random Forests. We work on a new auction dataset. This dataset contains real data on 1,292 auctions that transacted on eBay.com during May-June 2004. Suppose you are leading the analytics team at eBay and are tasked with predicting what types of auctions are likely to be competitive so that the platform can allocate more resources (e.g., advertising, supports) towards them. A competitive auction is defined as an auction with at least two bids placed on the item being auctioned. The data include information about the item (product category), the seller (number of ratings a seller has received), and the auction terms set by the seller (auction duration, opening price, whether the auction closing day is on weekend). You would like to build random forests to predict whether an auction will be competitive or not.